@@ -1,3 +1,5 @@
+//go:build !ignore_autogenerated
+
 /*
 Copyright 2024 chideat.
 
@@ -13,6 +15,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+
+// Code generated by controller-gen. DO NOT EDIT.
 
 package v1alpha1
 
@@ -99,26 +103,6 @@ func (in *ValkeyList) DeepCopyObject() runtime.Object {
 }
 
 // DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
-func (in *ValkeyModule) DeepCopyInto(out *ValkeyModule) {
-	*out = *in
-	if in.Args != nil {
-		in, out := &in.Args, &out.Args
-		*out = make([]string, len(*in))
-		copy(*out, *in)
-	}
-}
-
-// DeepCopy is an autogenerated deepcopy function, copying the receiver, creating a new ValkeyModule.
-func (in *ValkeyModule) DeepCopy() *ValkeyModule {
-	if in == nil {
-		return nil
-	}
-	out := new(ValkeyModule)
-	in.DeepCopyInto(out)
-	return out
-}
-
-// DeepCopyInto is an autogenerated deepcopy function, copying the receiver, writing into out. in must be non-nil.
 func (in *ValkeyReplicas) DeepCopyInto(out *ValkeyReplicas) {
 	*out = *in
 	if in.ShardsConfig != nil {
@@ -152,11 +136,7 @@ func (in *ValkeySpec) DeepCopyInto(out *ValkeySpec) {
 		*out = new(ValkeyReplicas)
 		(*in).DeepCopyInto(*out)
 	}
-	if in.Resources != nil {
-		in, out := &in.Resources, &out.Resources
-		*out = new(v1.ResourceRequirements)
-		(*in).DeepCopyInto(*out)
-	}
+	in.Resources.DeepCopyInto(&out.Resources)
 	if in.CustomConfigs != nil {
 		in, out := &in.CustomConfigs, &out.CustomConfigs
 		*out = make(map[string]string, len(*in))
@@ -166,7 +146,7 @@ func (in *ValkeySpec) DeepCopyInto(out *ValkeySpec) {
 	}
 	if in.Modules != nil {
 		in, out := &in.Modules, &out.Modules
-		*out = make([]ValkeyModule, len(*in))
+		*out = make([]core.ValkeyModule, len(*in))
 		for i := range *in {
 			(*in)[i].DeepCopyInto(&(*out)[i])
 		}
@@ -245,15 +225,8 @@ func (in *ValkeyStatus) DeepCopyInto(out *ValkeyStatus) {
 			(*out)[key] = val
 		}
 	}
-	if in.ProxyMatchLabels != nil {
-		in, out := &in.ProxyMatchLabels, &out.ProxyMatchLabels
-		*out = make(map[string]string, len(*in))
-		for key, val := range *in {
-			(*out)[key] = val
-		}
-	}
-	if in.ClusterNodes != nil {
-		in, out := &in.ClusterNodes, &out.ClusterNodes
+	if in.Nodes != nil {
+		in, out := &in.Nodes, &out.Nodes
 		*out = make([]core.ValkeyNode, len(*in))
 		copy(*out, *in)
 	}
