@@ -1,3 +1,5 @@
+//go:build !ignore_autogenerated
+
 /*
 Copyright 2024 chideat.
 
@@ -13,6 +15,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
+
+// Code generated by controller-gen. DO NOT EDIT.
 
 package v1alpha1
 
@@ -184,11 +188,7 @@ func (in *ClusterSpec) DeepCopyInto(out *ClusterSpec) {
 			(*out)[key] = val
 		}
 	}
-	if in.Resources != nil {
-		in, out := &in.Resources, &out.Resources
-		*out = new(v1.ResourceRequirements)
-		(*in).DeepCopyInto(*out)
-	}
+	in.Resources.DeepCopyInto(&out.Resources)
 	in.Access.DeepCopyInto(&out.Access)
 	if in.Storage != nil {
 		in, out := &in.Storage, &out.Storage
@@ -235,6 +235,13 @@ func (in *ClusterSpec) DeepCopyInto(out *ClusterSpec) {
 		in, out := &in.SecurityContext, &out.SecurityContext
 		*out = new(v1.PodSecurityContext)
 		(*in).DeepCopyInto(*out)
+	}
+	if in.Modules != nil {
+		in, out := &in.Modules, &out.Modules
+		*out = make([]core.ValkeyModule, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
 	}
 }
 
@@ -400,6 +407,13 @@ func (in *FailoverSpec) DeepCopyInto(out *FailoverSpec) {
 		in, out := &in.Sentinel, &out.Sentinel
 		*out = new(SentinelSettings)
 		(*in).DeepCopyInto(*out)
+	}
+	if in.Modules != nil {
+		in, out := &in.Modules, &out.Modules
+		*out = make([]core.ValkeyModule, len(*in))
+		for i := range *in {
+			(*in)[i].DeepCopyInto(&(*out)[i])
+		}
 	}
 }
 
