@@ -263,7 +263,7 @@ func buildPersistentClaims(cluster *v1alpha1.Cluster, labels map[string]string) 
 }
 
 func buildValkeyServerContainer(cluster *v1alpha1.Cluster, u *user.User, envs []corev1.EnvVar, index int) corev1.Container {
-	startArgs := []string{"sh", "/opt/run.sh"}
+	startArgs := []string{"sh", "/opt/run_cluster.sh"}
 
 	container := corev1.Container{
 		Name:            builder.ServerContainerName,
@@ -316,11 +316,11 @@ func buildValkeyServerContainer(cluster *v1alpha1.Cluster, u *user.User, envs []
 				},
 			},
 		},
-		Resources: *cluster.Spec.Resources,
+		Resources: cluster.Spec.Resources,
 		Lifecycle: &corev1.Lifecycle{
 			PreStop: &corev1.LifecycleHandler{
 				Exec: &corev1.ExecAction{
-					Command: []string{"sh", "-inst", "/opt/valkey-helper cluster shutdown  &> /proc/1/fd/1"},
+					Command: []string{"sh", "-c", "/opt/valkey-helper cluster shutdown  &> /proc/1/fd/1"},
 				},
 			},
 		},
