@@ -18,8 +18,8 @@ package actor
 
 import (
 	"context"
+	"maps"
 	"reflect"
-	"time"
 
 	"github.com/Masterminds/semver/v3"
 	"github.com/chideat/valkey-operator/api/core"
@@ -69,10 +69,11 @@ func (a *actorUpdateAccount) Do(ctx context.Context, val types.Instance) *actor.
 	logger := val.Logger().WithValues("actor", ops.CommandUpdateAccount.String())
 
 	var (
-		inst    = val.(types.FailoverInstance)
-		users   = inst.Users()
-		opUser  = users.GetOpUser()
-		ownRefs = util.BuildOwnerReferences(inst.Definition())
+		inst      = val.(types.FailoverInstance)
+		users     = inst.Users()
+		opUser    = users.GetOpUser()
+		ownRefs   = util.BuildOwnerReferences(inst.Definition())
+		isUpdated = false
 	)
 
 	name := aclbuilder.GenerateACLConfigMapName(inst.Arch(), inst.GetName())
@@ -80,7 +81,10 @@ func (a *actorUpdateAccount) Do(ctx context.Context, val types.Instance) *actor.
 	if err != nil && !errors.IsNotFound(err) {
 		logger.Error(err, "load configmap failed", "target", name)
 		return actor.NewResultWithError(ops.CommandRequeue, err)
-	} else if oldCm == nil {
+	} else if oldCm != nil && oldCm.GetDeletionTimestamp() != nil {
+		logger.Info("configmap is being deleted, skip update", "target", name)
+		return actor.NewResult(ops.CommandRequeue)
+	} else {
 		// sync acl configmap
 		oldCm = &corev1.ConfigMap{
 			ObjectMeta: metav1.ObjectMeta{
@@ -91,26 +95,28 @@ func (a *actorUpdateAccount) Do(ctx context.Context, val types.Instance) *actor.
 			},
 			Data: users.Encode(true),
 		}
+		isUpdated = true
 
-		// create acl with old password
-		// create acl file, after restart, the password is updated
-		if err := a.client.CreateConfigMap(ctx, inst.GetNamespace(), oldCm); err != nil {
-			logger.Error(err, "create acl configmap failed", "target", oldCm.Name)
-			return actor.NewResultWithError(ops.CommandRequeue, err)
-		}
+		// // create acl with old password
+		// // create acl file, after restart, the password is updated
+		// if err := a.client.CreateConfigMap(ctx, inst.GetNamespace(), oldCm); err != nil {
+		// 	logger.Error(err, "create acl configmap failed", "target", oldCm.Name)
+		// 	return actor.NewResultWithError(ops.CommandRequeue, err)
+		// }
 
-		// wait for resource sync
-		time.Sleep(time.Second * 1)
-		if oldCm, err = a.client.GetConfigMap(ctx, inst.GetNamespace(), name); err != nil {
-			logger.Error(err, "get configmap failed", "target", name)
-			return actor.NewResultWithError(ops.CommandRequeue, err)
-		}
+		// // wait for resource sync
+		// time.Sleep(time.Second * 1)
+		// if oldCm, err = a.client.GetConfigMap(ctx, inst.GetNamespace(), name); err != nil {
+		// 	logger.Error(err, "get configmap failed", "target", name)
+		// 	return actor.NewResultWithError(ops.CommandRequeue, err)
+		// }
 	}
 
-	isUpdated := false
+	logger.Info("update account", "namespace", inst.GetNamespace(), "name", inst.GetName(), "aclConfigMap", oldCm.Name, "opUser", opUser == nil)
+
 	if opUser == nil {
 		secretName := aclbuilder.GenerateACLOperatorSecretName(inst.Arch(), inst.GetName())
-		opUser, err := acl.NewOperatorUser(ctx, a.client, secretName, inst.GetNamespace(), ownRefs)
+		opUser, err = acl.NewOperatorUser(ctx, a.client, secretName, inst.GetNamespace(), ownRefs)
 		if err != nil {
 			logger.Error(err, "create operator user failed")
 			return actor.NewResult(ops.CommandRequeue)
@@ -131,10 +137,7 @@ func (a *actorUpdateAccount) Do(ctx context.Context, val types.Instance) *actor.
 		isUpdated = true
 	}
 	if isUpdated {
-		for k, v := range users.Encode(true) {
-			oldCm.Data[k] = v
-		}
-
+		maps.Copy(oldCm.Data, users.Encode(true))
 		if err := a.client.CreateOrUpdateConfigMap(ctx, inst.GetNamespace(), oldCm); err != nil {
 			logger.Error(err, "update acl configmap failed", "target", oldCm.Name)
 			return actor.NewResultWithError(ops.CommandRequeue, err)
