@@ -31,7 +31,6 @@ import (
 	"github.com/chideat/valkey-operator/internal/actor"
 	"github.com/chideat/valkey-operator/internal/builder"
 	"github.com/chideat/valkey-operator/internal/builder/aclbuilder"
-	"github.com/chideat/valkey-operator/internal/builder/clusterbuilder"
 	"github.com/chideat/valkey-operator/internal/builder/failoverbuilder"
 	"github.com/chideat/valkey-operator/internal/util"
 	"github.com/chideat/valkey-operator/internal/valkey/failover/monitor"
@@ -83,9 +82,9 @@ func (g *RuleEngine) Inspect(ctx context.Context, val types.Instance) *actor.Act
 	}
 
 	// NOTE: checked if resource is fullfilled
-	if isFullfilled, _ := inst.IsResourceFullfilled(ctx); !isFullfilled {
-		return actor.NewResult(CommandEnsureResource)
-	}
+	// if isFullfilled, _ := inst.IsResourceFullfilled(ctx); !isFullfilled {
+	// 	return actor.NewResult(CommandEnsureResource)
+	// }
 
 	// check password
 	if ret := g.isPasswordChanged(ctx, inst, logger); ret != nil {
@@ -150,7 +149,7 @@ func (g *RuleEngine) isPatchLabelNeeded(ctx context.Context, inst types.Failover
 		labelVal := labels[builder.RoleLabelKey]
 		if node == nil {
 			if labelVal != "" {
-				logger.V(3).Info("node not accessable", "name", pod.GetName(), "labels", labels)
+				logger.V(3).Info("node not accessible", "name", pod.GetName(), "labels", labels)
 				return actor.NewResult(CommandPatchLabels)
 			}
 			continue
@@ -201,8 +200,8 @@ func (g *RuleEngine) isConfigChanged(ctx context.Context, inst types.FailoverIns
 	} else if err != nil {
 		return actor.RequeueWithError(err)
 	}
-	newConf, _ := clusterbuilder.LoadValkeyConfig(newCm.Data[builder.ValkeyConfigKey])
-	oldConf, _ := clusterbuilder.LoadValkeyConfig(oldCm.Data[builder.ValkeyConfigKey])
+	newConf, _ := builder.LoadValkeyConfig(newCm.Data[builder.ValkeyConfigKey])
+	oldConf, _ := builder.LoadValkeyConfig(oldCm.Data[builder.ValkeyConfigKey])
 	added, changed, deleted := oldConf.Diff(newConf)
 	if len(added)+len(changed)+len(deleted) != 0 {
 		return actor.NewResult(CommandUpdateConfig)
