@@ -44,15 +44,19 @@ import (
 
 // nolint:unused
 // log is for logging in this package.
-var logger = logf.Log.WithName("user-resource")
+var logger = logf.Log.WithName("user-webhook")
 
 // SetupUserWebhookWithManager registers the webhook for User in the manager.
 func SetupUserWebhookWithManager(mgr ctrl.Manager) error {
 	return ctrl.NewWebhookManagedBy(mgr).For(&valkeybufredv1alpha1.User{}).
-		WithValidator(&UserCustomValidator{}).
-		WithDefaulter(&UserCustomDefaulter{}).
+		WithValidator(&UserCustomValidator{mgrClient: mgr.GetClient()}).
+		WithDefaulter(&UserCustomDefaulter{mgrClient: mgr.GetClient()}).
 		Complete()
 }
+
+const (
+	actionKey = "action"
+)
 
 // +kubebuilder:webhook:path=/mutate-valkey-buf-red-v1alpha1-user,mutating=true,failurePolicy=fail,sideEffects=None,groups=valkey.buf.red,resources=users,verbs=create;update;delete,versions=v1alpha1,name=muser-v1alpha1.kb.io,admissionReviewVersions=v1
 
@@ -69,12 +73,22 @@ var _ webhook.CustomDefaulter = &UserCustomDefaulter{}
 
 func (d *UserCustomDefaulter) Default(ctx context.Context, obj runtime.Object) (err error) {
 	inst, ok := obj.(*valkeybufredv1alpha1.User)
-
 	if !ok {
-		return fmt.Errorf("expected an User object but got %T", obj)
+		return fmt.Errorf("expected User object but got %T", obj)
+	}
+
+	{
+		var oldInst valkeybufredv1alpha1.User
+		if err := d.mgrClient.Get(ctx, client.ObjectKeyFromObject(inst), &oldInst); err != nil && !errors.IsNotFound(err) {
+			return err
+		}
+		if oldInst.GetDeletionTimestamp() != nil {
+			return nil
+		}
 	}
 
 	logger.V(3).Info("setup defaults", "user.name", inst.Name)
+
 	if inst.Labels == nil {
 		inst.Labels = make(map[string]string)
 	}
@@ -144,13 +158,12 @@ type UserCustomValidator struct {
 
 var _ webhook.CustomValidator = &UserCustomValidator{}
 
-// ValidateCreate implements webhook.CustomValidator so a webhook will be registered for the type User.
-func (v *UserCustomValidator) ValidateCreate(ctx context.Context, obj runtime.Object) (warns admission.Warnings, err error) {
-	inst, ok := obj.(*valkeybufredv1alpha1.User)
+func (v *UserCustomValidator) validate(ctx context.Context, oldObj, newObj runtime.Object) (warns admission.Warnings, err error) {
+	inst, ok := newObj.(*valkeybufredv1alpha1.User)
 	if !ok {
-		return nil, fmt.Errorf("expected a User object but got %T", obj)
+		return nil, fmt.Errorf("expected User object but got %T", newObj)
 	}
-	logger.Info("Validation for User upon creation", "name", inst.GetName())
+	logger.V(3).Info("Validation for User upon creation", "name", inst.GetName())
 
 	rule, err := user.NewRule(strings.ToLower(inst.Spec.AclRules))
 	if err != nil {
@@ -189,113 +202,51 @@ func (v *UserCustomValidator) ValidateCreate(ctx context.Context, obj runtime.Ob
 		}
 	}
 
-	switch inst.Spec.Arch {
-	case core.ValkeyFailover, core.ValkeyReplica:
-		rf := &v1alpha1.Failover{}
-		if err := v.mgrClient.Get(context.Background(), types.NamespacedName{
-			Namespace: inst.Namespace,
-			Name:      inst.Spec.InstanceName}, rf); err != nil {
-			return nil, err
-		}
-		if rf.Status.Phase != v1alpha1.FailoverPhaseReady {
-			return nil, fmt.Errorf("failover %s is not ready", inst.Spec.InstanceName)
-		}
-	case core.ValkeyCluster:
-		cluster := v1alpha1.Cluster{}
-		if err := v.mgrClient.Get(context.Background(), types.NamespacedName{
-			Namespace: inst.Namespace,
-			Name:      inst.Spec.InstanceName}, &cluster); err != nil {
-			return nil, err
-		}
-		if cluster.Status.Phase != v1alpha1.ClusterPhaseReady {
-			return nil, fmt.Errorf("cluster %s is not ready", inst.Spec.InstanceName)
+	if oldObj == nil {
+		switch inst.Spec.Arch {
+		case core.ValkeyFailover, core.ValkeyReplica:
+			vf := &v1alpha1.Failover{}
+			if err := v.mgrClient.Get(context.Background(), types.NamespacedName{
+				Namespace: inst.Namespace,
+				Name:      inst.Spec.InstanceName,
+			}, vf); err != nil {
+				return nil, err
+			}
+		case core.ValkeyCluster:
+			cluster := v1alpha1.Cluster{}
+			if err := v.mgrClient.Get(context.TODO(), types.NamespacedName{
+				Namespace: inst.Namespace,
+				Name:      inst.Spec.InstanceName,
+			}, &cluster); err != nil {
+				return nil, err
+			}
 		}
 	}
 	return
 }
 
+// ValidateCreate implements webhook.CustomValidator so a webhook will be registered for the type User.
+func (v *UserCustomValidator) ValidateCreate(ctx context.Context, obj runtime.Object) (warns admission.Warnings, err error) {
+	ctx = context.WithValue(ctx, actionKey, "create")
+	return v.validate(ctx, nil, obj)
+}
+
 // ValidateUpdate implements webhook.CustomValidator so a webhook will be registered for the type User.
 func (v *UserCustomValidator) ValidateUpdate(ctx context.Context, oldObj, newObj runtime.Object) (warns admission.Warnings, err error) {
-	inst, ok := newObj.(*valkeybufredv1alpha1.User)
-	if !ok {
-		return nil, fmt.Errorf("expected a User object for the newObj but got %T", newObj)
-	}
-	logger.Info("Validation for User upon update", "name", inst.GetName())
-
-	rule, err := user.NewRule(strings.ToLower(inst.Spec.AclRules))
-	if err != nil {
-		return nil, err
-	}
-
-	if inst.Spec.AccountType == v1alpha1.SystemAccount {
-		if inst.Spec.Username != user.DefaultOperatorUserName {
-			return nil, fmt.Errorf("system account username must be operator")
-		}
-		if rule.IsACLCommandsDisabled() {
-			return nil, fmt.Errorf("system account must enable acl command")
-		}
-		return
-	}
-
-	if inst.GetDeletionTimestamp() != nil {
-		return nil, nil
-	}
-	if len(inst.Spec.PasswordSecrets) == 0 {
-		return nil, fmt.Errorf("password secret can not be empty")
-	}
-
-	if err := rule.Validate(true); err != nil {
-		return nil, err
-	}
-	for _, passwordSecret := range inst.Spec.PasswordSecrets {
-		if passwordSecret == "" {
-			return nil, fmt.Errorf("password secret can not be empty")
-		}
-		secret := &v1.Secret{}
-		if err := v.mgrClient.Get(context.TODO(), types.NamespacedName{
-			Namespace: inst.Namespace,
-			Name:      passwordSecret,
-		}, secret); err != nil {
-			return nil, err
-		} else if err := security.PasswordValidate(string(secret.Data["password"]), 8, 32); err != nil {
-			return nil, err
-		}
-	}
-
-	switch inst.Spec.Arch {
-	case core.ValkeyFailover, core.ValkeyReplica:
-		vf := &v1alpha1.Failover{}
-		if err := v.mgrClient.Get(context.Background(), types.NamespacedName{
-			Namespace: inst.Namespace,
-			Name:      inst.Spec.InstanceName}, vf); err != nil {
-			return nil, err
-		}
-		if vf.Status.Phase != v1alpha1.FailoverPhaseReady {
-			return nil, fmt.Errorf("failover %s is not ready", inst.Spec.InstanceName)
-		}
-	case core.ValkeyCluster:
-		cluster := v1alpha1.Cluster{}
-		if err := v.mgrClient.Get(context.TODO(), types.NamespacedName{
-			Namespace: inst.Namespace,
-			Name:      inst.Spec.InstanceName}, &cluster); err != nil {
-			return nil, err
-		}
-		if cluster.Status.Phase != v1alpha1.ClusterPhaseReady {
-			return nil, fmt.Errorf("cluster %s is not ready", inst.Spec.InstanceName)
-		}
-	}
-	return
+	ctx = context.WithValue(ctx, actionKey, "update")
+	return v.validate(ctx, oldObj, newObj)
 }
 
 // ValidateDelete implements webhook.CustomValidator so a webhook will be registered for the type User.
 func (v *UserCustomValidator) ValidateDelete(ctx context.Context, obj runtime.Object) (warns admission.Warnings, err error) {
 	inst, ok := obj.(*valkeybufredv1alpha1.User)
 	if !ok {
-		return nil, fmt.Errorf("expected a User object but got %T", obj)
+		return nil, fmt.Errorf("expected User object but got %T", obj)
 	}
-	logger.Info("Validation for User upon deletion", "name", inst.GetName())
+	logger.V(3).Info("Validation for User upon deletion", "name", inst.GetName())
 
-	if inst.Spec.Username == user.DefaultUserName || inst.Spec.Username == user.DefaultOperatorUserName {
+	// TODO: added support of safely deleting default user
+	if inst.Spec.Username == user.DefaultOperatorUserName || inst.Spec.Username == user.DefaultUserName {
 		switch inst.Spec.Arch {
 		case core.ValkeyFailover, core.ValkeyReplica:
 			rf := v1alpha1.Failover{}
