@@ -51,6 +51,62 @@ func TestParseNodeFromClusterNode(t *testing.T) {
 			},
 			wantErr: false,
 		},
+		{
+			name: "with aux fields",
+			args: args{line: "343eca6406a48f95388682965bc19e37fcd300ca 192.168.130.82:31397@32669,nodename=node1,tcp-port=6379,tls-port=0,shard-id=a300430f8783c8a604dede6b8afabdc36230f038 master - 0 1743648449265 2 connected 5462-10922"},
+			want: &ClusterNode{
+				Id:      "343eca6406a48f95388682965bc19e37fcd300ca",
+				Addr:    "192.168.130.82:31397",
+				RawFlag: "master",
+				BusPort: "32669",
+				AuxFields: ClusterNodeAuxFields{
+					ShardID:  "a300430f8783c8a604dede6b8afabdc36230f038",
+					NodeName: "node1",
+					TCPPort:  6379,
+					TLSPort:  0,
+					raw:      "192.168.130.82:31397@32669,nodename=node1,tcp-port=6379,tls-port=0,shard-id=a300430f8783c8a604dede6b8afabdc36230f038",
+				},
+				Role:      "master",
+				MasterId:  "",
+				PingSend:  0,
+				PongRecv:  1743648449265,
+				Epoch:     2,
+				LinkState: "connected",
+				slots:     []string{"5462-10922"},
+				rawInfo:   "343eca6406a48f95388682965bc19e37fcd300ca 192.168.130.82:31397@32669,nodename=node1,tcp-port=6379,tls-port=0,shard-id=a300430f8783c8a604dede6b8afabdc36230f038 master - 0 1743648449265 2 connected 5462-10922",
+			},
+			wantErr: false,
+		},
+		{
+			name: "with invalid aux fields",
+			args: args{line: "343eca6406a48f95388682965bc19e37fcd300ca 192.168.130.82:31397@32669,node1,tcp-port=6379,tls-port=0,shard-id=a300430f8783c8a604dede6b8afabdc36230f038 master - 0 1743648449265 2 connected 5462-10922"},
+			want: &ClusterNode{
+				Id:      "343eca6406a48f95388682965bc19e37fcd300ca",
+				Addr:    "192.168.130.82:31397",
+				RawFlag: "master",
+				BusPort: "32669",
+				AuxFields: ClusterNodeAuxFields{
+					ShardID: "a300430f8783c8a604dede6b8afabdc36230f038",
+					TCPPort: 6379,
+					TLSPort: 0,
+					raw:     "192.168.130.82:31397@32669,node1,tcp-port=6379,tls-port=0,shard-id=a300430f8783c8a604dede6b8afabdc36230f038",
+				},
+				Role:      "master",
+				MasterId:  "",
+				PingSend:  0,
+				PongRecv:  1743648449265,
+				Epoch:     2,
+				LinkState: "connected",
+				slots:     []string{"5462-10922"},
+				rawInfo:   "343eca6406a48f95388682965bc19e37fcd300ca 192.168.130.82:31397@32669,node1,tcp-port=6379,tls-port=0,shard-id=a300430f8783c8a604dede6b8afabdc36230f038 master - 0 1743648449265 2 connected 5462-10922",
+			},
+			wantErr: false,
+		},
+		{
+			name:    "with error addr",
+			args:    args{line: "343eca6406a48f95388682965bc19e37fcd300ca 192.168.130.82:31397,nodename=node1,tcp-port=6379,tls-port=0,shard-id=a300430f8783c8a604dede6b8afabdc36230f038 master - 0 1743648449265 2 connected 5462-10922"},
+			wantErr: true,
+		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
@@ -61,6 +117,10 @@ func TestParseNodeFromClusterNode(t *testing.T) {
 			}
 			if !reflect.DeepEqual(got, tt.want) {
 				t.Errorf("ParseNodeFromClusterNode() = %v, want %v", got, tt.want)
+				return
+			}
+			if !tt.wantErr && got.AuxFields.Raw() != tt.want.AuxFields.Raw() {
+				t.Errorf("ClusterNodeAuxFields.Raw() = %v, want %v", got.AuxFields.Raw(), tt.want.AuxFields.Raw())
 			}
 		})
 	}
@@ -69,12 +129,12 @@ func TestParseNodeFromClusterNode(t *testing.T) {
 func TestClusterNode_IsSelf(t *testing.T) {
 	tests := []struct {
 		name   string
-		fields ClusterNode
+		fields *ClusterNode
 		want   bool
 	}{
 		{
 			name: "isSelf",
-			fields: ClusterNode{
+			fields: &ClusterNode{
 				Id:        "33b1262d41a4d9c27a78eef522c84999b064ce7f",
 				Addr:      "",
 				RawFlag:   "myself,master",
@@ -86,20 +146,28 @@ func TestClusterNode_IsSelf(t *testing.T) {
 			},
 			want: true,
 		},
+		{
+			name:   "nil",
+			fields: nil,
+			want:   false,
+		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			n := &ClusterNode{
-				Id:        tt.fields.Id,
-				Addr:      tt.fields.Addr,
-				RawFlag:   tt.fields.RawFlag,
-				MasterId:  tt.fields.MasterId,
-				PingSend:  tt.fields.PingSend,
-				PongRecv:  tt.fields.PongRecv,
-				Epoch:     tt.fields.Epoch,
-				LinkState: tt.fields.LinkState,
-				slots:     tt.fields.slots,
-				Role:      tt.fields.Role,
+			var n *ClusterNode
+			if tt.fields != nil {
+				n = &ClusterNode{
+					Id:        tt.fields.Id,
+					Addr:      tt.fields.Addr,
+					RawFlag:   tt.fields.RawFlag,
+					MasterId:  tt.fields.MasterId,
+					PingSend:  tt.fields.PingSend,
+					PongRecv:  tt.fields.PongRecv,
+					Epoch:     tt.fields.Epoch,
+					LinkState: tt.fields.LinkState,
+					slots:     tt.fields.slots,
+					Role:      tt.fields.Role,
+				}
 			}
 			if got := n.IsSelf(); got != tt.want {
 				t.Errorf("ClusterNode.IsSelf() = %v, want %v", got, tt.want)
@@ -128,6 +196,22 @@ func TestClusterNodes_Self(t *testing.T) {
 			name: "self",
 			ns:   []*ClusterNode{&node},
 			want: &node,
+		},
+		{
+			name: "no self",
+			ns: []*ClusterNode{
+				{
+					Id:        "33b1262d41a4d9c27a78eef522c84999b064ce7f",
+					Addr:      "",
+					RawFlag:   "master",
+					MasterId:  "",
+					PingSend:  0,
+					PongRecv:  0,
+					Epoch:     0,
+					LinkState: "connected",
+				},
+			},
+			want: nil,
 		},
 	}
 	for _, tt := range tests {
@@ -524,22 +608,52 @@ func TestParseNodes(t *testing.T) {
 		wantErr bool
 	}{
 		{
-			name: "valid nodes",
+			name: "valid nodes without host",
 			data: `33b1262d41a4d9c27a78eef522c84999b064ce7f :6379@16379 myself,master - 0 0 0 connected
-44c1262d41a4d9c27a78eef522c84999b064ce7f :6380@16380 slave 33b1262d41a4d9c27a78eef522c84999b064ce7f 0 0 0 connected`,
+44c1262d41a4d9c27a78eef522c84999b064ce7f :6380@16380 slave 33b1262d41a4d9c27a78eef522c84999b064ce7f 0 0 0 connected
+vars currentEpoch 5 lastVoteEpoch 0`,
 			want:    2,
+			wantErr: false,
+		},
+		{
+			name: "valid nodes",
+			data: `b4cb5b221aedca8b7380f51bc3a10195e0211066 192.168.138.234:30643@31899,,tls-port=0,shard-id=48a6e32a9da7c8d484ff932ce0bb9f540a930ed9 slave ba1577ec8d5037199b4e9fe7661ae569995713d9 0 1748928991000 9872 connected
+ba1577ec8d5037199b4e9fe7661ae569995713d9 192.168.142.242:32412@30910,,tls-port=0,shard-id=48a6e32a9da7c8d484ff932ce0bb9f540a930ed9 master - 0 1748928992000 9872 connected 0-5461
+c71c6a6de4e3309e40669720863da816d0d28835 192.168.142.242:30689@30476,,tls-port=0,shard-id=ad7120bc83b9491e179c79b747ca60c610285648 slave ffc9f22c82a69d6d56699513ac5a2eb55a1ab267 0 1748928993000 9864 connected
+6a8a3f4037ce81183cf2f0dbade7422f4a553514 192.168.142.242:31195@32311,,tls-port=0,shard-id=d435d01b17038215998f66759e1800b977453e17 slave d386323275dfce0e0b8c4c645a20c5892fa47566 0 1748928993246 9865 connected
+d386323275dfce0e0b8c4c645a20c5892fa47566 192.168.138.234:31012@30749,,tls-port=0,shard-id=d435d01b17038215998f66759e1800b977453e17 master - 0 1748928992186 9865 connected 10923-16383
+ffc9f22c82a69d6d56699513ac5a2eb55a1ab267 192.168.138.234:32320@30296,,tls-port=0,shard-id=ad7120bc83b9491e179c79b747ca60c610285648 myself,master - 0 1748928992000 9864 connected 5462-10922
+vars currentEpoch 9872 lastVoteEpoch 9872`,
+			want:    6,
+			wantErr: false,
+		},
+		{
+			name: "valid cluster nodes",
+			data: `b4cb5b221aedca8b7380f51bc3a10195e0211066 192.168.138.234:30643@31899 slave ba1577ec8d5037199b4e9fe7661ae569995713d9 0 1748929224000 9872 connected
+ba1577ec8d5037199b4e9fe7661ae569995713d9 192.168.142.242:32412@30910 master - 0 1748929226515 9872 connected 0-5461
+c71c6a6de4e3309e40669720863da816d0d28835 192.168.142.242:30689@30476 slave ffc9f22c82a69d6d56699513ac5a2eb55a1ab267 0 1748929225512 9864 connected
+6a8a3f4037ce81183cf2f0dbade7422f4a553514 192.168.142.242:31195@32311 slave d386323275dfce0e0b8c4c645a20c5892fa47566 0 1748929225000 9865 connected
+d386323275dfce0e0b8c4c645a20c5892fa47566 192.168.138.234:31012@30749 master - 0 1748929223000 9865 connected 10923-16383
+ffc9f22c82a69d6d56699513ac5a2eb55a1ab267 192.168.138.234:32320@30296 myself,master - 0 1748929225000 9864 connected 5462-10922`,
+			want:    6,
 			wantErr: false,
 		},
 		{
 			name:    "empty data",
 			data:    "",
 			want:    0,
-			wantErr: true,
+			wantErr: false,
 		},
 		{
 			name:    "invalid node format",
 			data:    "invalid node data",
 			wantErr: true,
+		},
+		{
+			name:    "cluster nodes test",
+			data:    "b4cb5b221aedca8b7380f51bc3a10195e0211066 192.168.138.234:30643@31899 slave ba1577ec8d5037199b4e9fe7661ae569995713d9 0 1748929519136 9872 connected\nba1577ec8d5037199b4e9fe7661ae569995713d9 192.168.142.242:32412@30910 master - 0 1748929520141 9872 connected 0-5461\nc71c6a6de4e3309e40669720863da816d0d28835 192.168.142.242:30689@30476 slave ffc9f22c82a69d6d56699513ac5a2eb55a1ab267 0 1748929519000 9864 connected\n6a8a3f4037ce81183cf2f0dbade7422f4a553514 192.168.142.242:31195@32311 slave d386323275dfce0e0b8c4c645a20c5892fa47566 0 1748929521146 9865 connected\nd386323275dfce0e0b8c4c645a20c5892fa47566 192.168.138.234:31012@30749 master - 0 1748929519000 9865 connected 10923-16383\nffc9f22c82a69d6d56699513ac5a2eb55a1ab267 192.168.138.234:32320@30296 myself,master - 0 1748929518000 9864 connected 5462-10922\n",
+			want:    6,
+			wantErr: false,
 		},
 	}
 	for _, tt := range tests {
